deps - replace styled-components with goober
